requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
colorama==0.4.6
Django==4.2.3
Pillow==10.0.0
pydantic==2.0.3
pydantic_core==2.3.0
pypng==0.20220715.0
pytz==2023.3
qrcode==7.4.2
segno==1.5.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
